feat: 다음페이지 이동할 때 정렬에 사용하는 필드 입력받도록 변경

+++ coffice-domain/src/main/java/kr/co/yapp/_22nd/coffice/domain/place/PlaceQueryService.java
@@ -15,7 +15,7 @@ public interface PlaceQueryService {
 
     Place getPlace(Long placeId);
 
-    Slice<PlaceSearchResponseVo> search(PlaceSearchRequestVo placeSearchVo, CursorPageable<Long> cursorPageable);
+    Slice<PlaceSearchResponseVo> search(PlaceSearchRequestVo placeSearchVo, CursorPageable<Double> cursorPageable);
 
     List<Place> findByPlaceFolderId(Long memberId, Long placeFolderId);
 }
